dataplex_integration/requirements.txt
requests
google-cloud-dataplex
google-cloud-bigquery
google-cloud-storage
google-cloud-pubsub
